// src/lib/oracles/bandProtocol/types.ts
export interface ValidatorInfo {
  operatorAddress: string;
  moniker: string;
  identity: string;
  website: string;
  details: string;
  tokens: number;
  delegatorShares: number;
  commissionRate: number;
  maxCommissionRate: number;
  maxCommissionChangeRate: number;
  uptime: number;
  jailed: boolean;
  rank: number;
}
// ...
export type HistoryPeriod = 7 | 30 | 90;